Moves nodes filepath and updated some links
