+++ src/route/nodes/Nodes.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'react-i18next'
 import { NodeTable } from './components/NodeTable'
 import { NetworkSelect } from '../../components/NetworkSelect'
 import { Search, Sync } from '@mui/icons-material'
-import { NetworkNodes } from '../networks/networkId/nodes/NetworkNodes'
+import { NetworkNodes } from './netid/NetworkNodes'
 
 export const Nodes: React.FC = () => {
   const { path } = useRouteMatch()
