+++ src/route/nodes/netid/NetworkNodes.tsx
@@ -10,11 +10,11 @@ import { NodeId } from './nodeId/NodeId'
 import { Chip, Grid, IconButton, InputAdornment, TextField, Tooltip, Typography } from '@mui/material'
 import { encode64 } from '~util/fields'
 import { AltRoute, CallMerge, CallSplit, Delete, Search, Sync } from '@mui/icons-material'
-import { i18n } from '../../../../i18n/i18n'
+import { i18n } from '../../../i18n/i18n'
 import { CreateEgress } from './components/CreateEgress'
 import { TableToggleButton } from './components/TableToggleButton'
 import { CreateRelay } from './components/CreateRelay'
-import { NetworkSelect } from '../../../../components/NetworkSelect'
+import { NetworkSelect } from '../../../components/NetworkSelect'
 import { useDispatch, useSelector } from 'react-redux'
 import { authSelectors } from '~store/selectors'
 import { deleteNode, getNodes } from '~store/modules/node/actions'
