+++ src/route/nodes/netid/nodeId/NodeId.tsx
